+++ filespec.c
@@ -35,7 +35,7 @@ void SetMatchingParam(DirEntry *dir_entry)
   DirEntry  *de_ptr;
   FileEntry *fe_ptr;
   unsigned long  matching_files;
-  LONGLONG matching_bytes;
+  long long matching_bytes;
 
   for( de_ptr = dir_entry; de_ptr; de_ptr = de_ptr->next )
   {
@@ -55,14 +55,14 @@ void SetMatchingParam(DirEntry *dir_entry)
 	fe_ptr->matching = FALSE;
       }
     }
-   
+
     de_ptr->matching_files = matching_files;
     de_ptr->matching_bytes = matching_bytes;
 
     statistic.disk_matching_files += matching_files;
     statistic.disk_matching_bytes += matching_bytes;
-    
-    if( de_ptr->sub_tree ) 
+
+    if( de_ptr->sub_tree )
     {
       SetMatchingParam( de_ptr->sub_tree );
     }
@@ -77,7 +77,7 @@ Take in the user-specified new filespec.
 As modified, it defaults to '*'; the original version offered the
 current value as default, but that's just an up-arrow away.
 Returns 0 on success, -1 on failure (empty string).
-<<***************************************************************/            
+<<***************************************************************/
 
 int ReadFileSpec(void)
 {
